feat: add methods for going back in time
SInce we need access to the Game object in order to playback the moves, we have to include it. But Game also includes TimeMachine, resulting to a circular dependency.
In order to avoid this, we forward declare TimeMachine on Game instead of including it. The linker will take care of the rest.